Add BLE starter code
This is simply the code from the BLE tutorial. All it does is allow nRF
connect or something similar to change the LED color showing on the
nano. 1,2,3 are the options and they correspond to red, green, blue.
This is the backwards communication from what we actually want, but this
is a starting point