[COMM] Initial BLE scanning
Add to CIQBLE file necessary Delegate class to handle call backs of
BLE related functions including scan results. Combine multiple scan
results into a single array that can be gotten through the getter.
Due to confusion around how to properly initialize the array for
this kind of work, the getter function removes the first element of
the array (the initialization 'null' value, returning only scanned
values.

Add a Scanner UI that displays a scanning message, triggers a scan
and after 5 seconds, will end the scan, gather the results and
display the device names. This page is still greatly flawed and
the following need to be fixed before scanning can be considered
complete:
- Scan lasts a minimum of 5 seconds but depends on the screen
  refreshing which may only occur once the screen goes to sleep,
  then wakes up again (which is inconvenient, although fnuctional).
- Scan results are only displayed as a the first few items and need
  to be turned into a selectable menu for users (maybe a Menu class
  or selectable or something else, unknown).

Additional problems with this state is that the displayed devices
all carry the same name. It is unclear if this is because that is
what is being scanned, and error in the display, or an error in
the logic.

The purpose of this commit is to save a relatively functioning
place and to ensure that future work can pick up from here.

Testing:
- Use the latest build and load it onto a Venu 2
- Open the CIQ application and allow to fall asleep
- Awake the device and see 5 devices names (likely identical names)
  displayed. Note that there will be an error if less than 5
  devices exist since this is not a proper set up.

Jira: https://athenamr.atlassian.net/browse/CAPSTONE-40