[COMM] Fix 5 second time out
Change the 5 second scan to prevent reliance on screen refresh and
instead use a 5 second timer to trigger a callback. Move the
scanning results to a new temporary UI class that can be pushed
onto the view stack (pushed instead of swiches so that the back
button returns to scanning).

Testing:
- Change watch display sleep timeout to "long"
- Open the application and wait 5 seconds to see a new page with
  obvious scan results (actual devices will vary based on location)
- Hit the back button and return to the scanning page
- Wait for the scan results page to show up again (results may vary
  from previous attempts).

Jira: https://athenamr.atlassian.net/browse/CAPSTONE-40