Example code to scan successfully
This code will display "Found it" on the watch face if the watch finds
the arduino from scanning. Note that this does not connect the watch and
arduino yet. Also note that if the device is connected to anything else
(like nRF connect) it will not be findable by the watch. Our previous
issue was relying on device name to find it from a scan, but this is a
known bug:
https://forums.garmin.com/developer/connect-iq/f/discussion/279281/ble-device-names-null

Testing and Verification:
  Arduino:
    1. Flash the arduino with peripheral_successful_scanning.ino (note
       that you can use either the regular nanos or the sense)
    2. Begin the serial moniter to start advertising
    3. When testing with the watch, ensure that the arduino is not
       connected to anything
  Watch:
    1. Build the CIQ app using the "Build for Device" command
    2. Place the prg file it creates into GARMIN/APPS on the watch
    3. When ready to test, tap on the CIQ app and wait for the scanning
       timeout to hit
    4. When the watch face turns off, turn it back on an you will see
       "Found it" if successful