[COMM] MVP Communication complete
Subscribe to notifications works without a runtime crash. Runtime
crash fixed by displaying the first value in the array received
from the BLE connection.

Clean up comments and code to be more concise.

Added a new "Connecting..." page for the time delay between
connecting to the arduino and subscribing to notifications. This
avoids an awkward home page where we are displaying sound of zero
while waiting for the BLE stuff to complete.

Modify sound threshold default to 80 which is more reasonable.

Adjust Sensor disconnected UI to properly display all on the screen.

Adjust sound notification to change from red to white and vice versa
only when the sound threshold is exceeded during the notification.
Current delay is 15 seconds or the back button for users.

Testing:
- Use the Latest build on the watch and Fix BLE Notifications on
  arduino branch.
- Open the app and see scanning for around 5 seconds.
- See connecting for a few seconds followed by a switch to the
  home page.
- Talk or scream into the arduino to see db changes.
- Scream into the arduino to trigger a threshold notification and
  see the page switch. When above 80, the current value will be red
  and when below, it will be white.
- Unplug the arduino to see a disconnection message.
- Set the arduino back to advertising to see a reconnection. If no
  reconnection occurs, there may have been a timeout delay.

Jira:
- https://athenamr.atlassian.net/browse/CAPSTONE-41
- https://athenamr.atlassian.net/browse/CAPSTONE-31
- https://athenamr.atlassian.net/browse/CAPSTONE-29
- https://athenamr.atlassian.net/browse/CAPSTONE-30