[UI] Fix broken back button
Back button would pop from stack and we were not properly adjusting
the stack to be set up properly. Add a back delegate that sets up
the page to go back to (or null to exit application) and make every
page be a switch instead of a push or a pop.

Testing:
- NOT TESTED IN THIS COMMIT
- Download with Arduino MVP code and repeat the MVP testing that
  caused the bug