[UI] Add settings landing page
Settings landing page is now accessible by swiping up from any
non-settings page. The non-settings page can be returned to by swiping
down on a settings page. This is the first step to changing threshold
values since we need to account for both sound and temp, which would be
available to chose from on this landing page

Testing and Verification Instructions:
 - Build run this code on the simulator
 - Ensure the swiping as outlined above allowd navigation between
   settings and non-settings pages