[UI] Add default to picker
Added a default value to the sound picker so that it starts with the
current value. Also added TODOs for when dev/temp is merged, and any
comment starting with ----- is what I think the code will be. Also fixed
getIndex so I could use it for defaults

Testing and Verification Instructions:
 - Run this code on the simulator
 - See that the sound picker has the 80 dB default
 - Change that value and go back into settings to see the new threshold
   as the default