[COMM] Fix multiple subscriptions
Introduce a delay after the first characteristic is subscribed to. The
requestWrite calls need to either have a delay between them or be queued
so that they don't cause a runtime crash. The forum that explains this
behavior can be found below:
https://forums.garmin.com/developer/connect-iq/f/discussion/258434/how-to-do-multiple-successive-btle-characteristic-requestwrite/1234530#1234530

This commit has the delay solution and will work correctly. However, the
queuing solution is better and we will be working on making that change.
For now this will unblock the rest of our development. Once the
connection has been acheived, the subscriptions will happen, and then
the values populate the screen in real time.

Note: This commit has notifications disabled to avoid unknown behavior

Testing and Verification Instructions:
 - Flash the watch with the CIQ.prg in Latest Build
 - Connect to the device
 - See both characteristics populate the screen and update in realtime