[COMM] Modify BLE subscriptions from delay to queue
Add a queue class that can support queuing of notification
subscriptions for BLE. Class accepts adding characteristics to the
queue. Run is then triggerred once where the chars were added to
queue (connection confirmation). Running of the rest of the queue
is triggered after an item is confirmed with a successful write.

Testing:
- Harry will test the functionality and any future fixes will
  occur in future commits.