[UI] Fix null temp threshold bug
This now allows either max or min temp to be null. Also add global
variables for defining our threshold options. Also make it so that the
max thresh cannot be lower than min and vice versa. For example, if
the max temp is 30, min can be 29 at most.

Testing Instructions:
 - Set up system with this build and connect
 - Change one of the temp thresholds
 - Ensure the range of the other temp threshold makes sense
 - Change one of the temp thresholds to OFF
 - Ensure the watch doesn't crash, and the other threshold can be
   anything from the temp range