[COMM] Shorten connection time
Use a repeating timer to check if a sensor has been found. Repeat
timers are necessary because non-repeat timers cannot be stopped.
Total scan time remains a full 5 seconds with check intervals of
500 ms. If the sensor is found before the full 5 seconds, the
timer will be stopped and the connection process will be started.
If the sensor is not found, the sensor not found page will be
shown.

Testing:
- Builds without errors
- Functionality remains the same except for the early connection