Add activity visual
On Garmin Connect, when the overlay option is selected, it does not use
the same axes. This means that when threshold values are plotted over
current values, the intersections mean nothing. To fix this, I
downloaded one of the activity's fit files, converted it to csv using
the FitToCsv tool from the fit sdk. I then cleaned out all the data we
don't care about it and got chatGPT to plot graphs. This whole process
is contained in the folder added in this commit